Add support for "unaligned" processing
Support unaligned FFHQ as used in the paper

  "Alias-Free Generative Adversarial Networks"

Add --random-shift/--retry-crops/--no-rotation/--no-padding options
for controlling alignment.

Add --source-dir argument that can be used to point to a locally
downloaded FFHQ release.  This way the script is able to use
in-the-wild source images from a local directory rather than
trying to pull them from Google Drive.